--- app/src/main/java/com/example/workspace/schedule/presentation/ViewModel.java
@@ -1,7 +1,0 @@
-package com.example.workspace.schedule.presentation;
-
-public interface ViewModel<T extends BaseView> {
-
-    void init(T view);
-
-}
