+++ app/src/main/java/com/example/workspace/schedule/presentation/home/HomeView.java
@@ -6,4 +6,5 @@ import com.example.workspace.schedule.presentation.BaseView;
 
 public interface HomeView extends BaseView {
 
+    void onClicked(View view);
 }
