--- app/src/main/java/com/example/workspace/schedule/presentation/home/plans/NewPlanViewModelImpl.java
@@ -1,17 +1,0 @@
-package com.example.workspace.schedule.presentation.home.plans;
-
-import android.view.View;
-
-public class NewPlanViewModelImpl implements NewPlanViewModel {
-    private NewPlanView planView;
-
-    @Override
-    public void init(NewPlanView view) {
-        this.planView = view;
-    }
-
-    @Override
-    public void onClick(View view) {
-
-    }
-}
